# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project 2 CRUD API with Auth",
    "version": "1.0.0",
    "description": "API documentation for Project 2 with JWT Authentication"
  },
  "host": "cse341-project2-1-n6h8.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer <token>'"
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "consumes": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "schema": {
            "properties": {
              "username": { "type": "string" },
              "password": { "type": "string" },
              "email": { "type": "string" }
            },
            "required": ["username", "password", "email"]
          }
        }],
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Invalid input or username exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user and get JWT token",
        "consumes": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "schema": {
            "properties": {
              "username": { "type": "string" },
              "password": { "type": "string" }
            },
            "required": ["username", "password"]
          }
        }],
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Invalid username or password" }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": { "200": { "description": "Success" } }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Create a new product (protected)",
        "consumes": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "schema": {
            "properties": {
              "name": { "type": "string" },
              "category": { "type": "string" },
              "description": { "type": "string" },
              "price": { "type": "number" },
              "stock": { "type": "number" },
              "brand": { "type": "string" },
              "sku": { "type": "string" }
            },
            "required": ["name", "category", "price"]
          }
        }],
        "responses": {
          "201": { "description": "Product created successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}